% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{boot}
\alias{boot}
\alias{boot.default}
\alias{boot.VarExp}
\title{Derive bootstrap confidence intervals for variance decomposition (based on percentile method)}
\usage{
boot(object, ...)

\method{boot}{default}(object, ...)

\method{boot}{VarExp}(object, ...)
}
\arguments{
\item{object}{a \code{VarExp} object from varianceExplained()}

\item{...}{arguments passed to  \code{\link[lme4:bootMer]{lme4::bootMer()}}, in particular \code{nsim} for the number of simulations, the type of bootstrap and arguments for parallel computing.}

\item{progress}{only mmer objects. passed to plyr::llply}
}
\description{
Derive bootstrap confidence intervals for variance decomposition (based on percentile method)
}
