% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootVarianceExplained}
\alias{bootVarianceExplained}
\alias{bootVarianceExplained.default}
\alias{bootVarianceExplained.lmerMod}
\title{Derive bootstrap confidence intervals for variance decomposition}
\usage{
bootVarianceExplained(object, ...)

\method{bootVarianceExplained}{default}(object, ...)

\method{bootVarianceExplained}{lmerMod}(object, ...)
}
\arguments{
\item{object}{a \code{lmerMod} or \code{lmerModLmerTest} object created by \code{\link[lme4:lmer]{lme4::lmer()}} or \code{\link[lmerTest:lmer]{lmerTest::lmer()}}, respectively, or a \code{mmer} object created by  \code{\link[sommer:mmer]{sommer::mmer()}}.}

\item{...}{arguments passed to  \code{\link[lme4:bootMer]{lme4::bootMer()}}, in particular \code{nsim} for the number of simulations, thy type of bootstrap and arguments for parallel computing}
}
\description{
Derive bootstrap confidence intervals for variance decomposition
}
