% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expandResults.R
\name{expandResults}
\alias{expandResults}
\alias{expandResults.summary.VarExpProp}
\title{Add explained variance to model summary}
\usage{
expandResults(object, digits, ...)

\method{expandResults}{summary.VarExpProp}(object, digits = 1, ...)
}
\arguments{
\item{object}{a \code{summary.VarExpProp} object after boostrapping}

\item{...}{Currently no additional arguments}
}
\description{
Provide description here
}
\details{
Results provide proportion of variance explained in percentages
}
\examples{
# provide some example code

}
