% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explainedCovariance.R
\name{explainedCovariance}
\alias{explainedCovariance}
\alias{explainedCovariance.VarExp}
\title{Extract explained covariance matrix by random effects}
\usage{
explainedCovariance(object, ...)

\method{explainedCovariance}{VarExp}(object, ...)
}
\arguments{
\item{object}{a \code{VarExp} object}

\item{...}{Currently no additional arguments}
}
\value{
a matrix
}
\description{
Extract explained covariance matrix by random effects
}
