% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary.VarExp}
\alias{print.summary.VarExp}
\alias{print.summary.VarExpProp}
\title{Print  summary}
\usage{
\method{print}{summary.VarExp}(x, ...)

\method{print}{summary.VarExpProp}(x, ...)
}
\arguments{
\item{x}{object created by summary()}

\item{...}{arguments passed to print()}
}
\description{
Print  summary
}
