% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.VarExp}
\alias{print.VarExp}
\alias{print.VarExpProp}
\title{Print}
\usage{
\method{print}{VarExp}(x, ...)

\method{print}{VarExpProp}(x, ...)
}
\arguments{
\item{x}{\code{VarExp} object created by \code{\link[varianceExplainedPack:varianceExplained]{varianceExplainedPack::varianceExplained()}}.or \code{varExpProp} object created by \code{\link[varianceExplainedPack:proportionOf]{varianceExplainedPack::proportionOf()}}}
}
\description{
Print
}
