% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2.R
\name{R2}
\alias{R2}
\alias{R2.default}
\alias{R2.lmerMod}
\title{Coefficient of Determination}
\usage{
R2(x)

\method{R2}{default}(x)

\method{R2}{lmerMod}(x)
}
\arguments{
\item{object}{a \code{lmerMod} or \code{lmerModLmerTest} object created by \code{\link[lme4:lmer]{lme4::lmer()}} or \code{\link[lmerTest:lmer]{lmerTest::lmer()}}, respectively, or a \code{mmer} object created by  \code{\link[sommer:mmer]{sommer::mmer()}}.}
}
\description{
Coefficient of Determination
}
