% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proportionOf.R
\name{proportionOf}
\alias{proportionOf}
\title{Derive proportion of variance explained by variance components}
\usage{
proportionOf(object)
}
\arguments{
\item{object}{\code{varExp} object created by \code{\link[varianceExplainedPack:varianceExplained]{varianceExplainedPack::varianceExplained()}}.}
}
\description{
Derive proportion of variance explained by variance components
}
