% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.VarExp}
\alias{summary.VarExp}
\alias{summary.VarExpProp}
\title{Summary for (proportion of) variance explained by variance components}
\usage{
\method{summary}{VarExp}(object, digits = 1, ...)

\method{summary}{VarExpProp}(object, ...)
}
\arguments{
\item{object}{\code{varExp} object created by \code{\link[varianceExplainedPack:varianceExplained]{varianceExplainedPack::varianceExplained()}}.or \code{varExpProp} object created by \code{\link[varianceExplainedPack:proportionOf]{varianceExplainedPack::proportionOf()}}}
}
\description{
Summary for (proportion of) variance explained by variance components
}
