% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varianceExplained.R
\name{varianceExplained}
\alias{varianceExplained}
\alias{varianceExplained.default}
\alias{varianceExplained.lmerMod}
\alias{varianceExplained.mmer}
\title{Derive variance decomposition}
\usage{
varianceExplained(object, ...)

\method{varianceExplained}{default}(object, ...)

\method{varianceExplained}{lmerMod}(object, cholesky = TRUE, ...)

\method{varianceExplained}{mmer}(object, X, Z, cholesky = TRUE, ...)
}
\arguments{
\item{object}{a \code{lmerMod} or \code{lmerModLmerTest} object created by \code{\link[lme4:lmer]{lme4::lmer()}} or \code{\link[lmerTest:lmer]{lmerTest::lmer()}}, respectively, or a \code{mmer} object created by  \code{\link[sommer:mmer]{sommer::mmer()}}.}

\item{...}{Currently no additional arguments}

\item{X}{Design matrix for fixed effects}

\item{Z}{\emph{named} List of desgin matrices for random effects (one list element for each effect).}
}
\value{
a varExp object
}
\description{
Provide description here
}
\details{
Provide details here
}
\examples{
# provide some example code

}
